chatbot/aws-bedrock/voice-chat/server/requirements.txt
# requirements.txt
fastapi
uvicorn
python-dotenv
boto3
botocore
httpx
pydantic
edge-tts
websockets